+++ mfhdf/nctest/vputgetg.c
@@ -120,22 +120,22 @@ test_varputgetg(cdfid)
 		errvar(&test, &test.vars[iv]);
 		(void)fprintf(stderr,"  corner = (");
 		for (id = 0 ; id < test.vars[iv].ndims; id++)
-		  (void)fprintf(stderr,"%d%s",hc[ie].cor[id],
+		  (void)fprintf(stderr,"%d%s",(int)hc[ie].cor[id],
 				(id < test.vars[iv].ndims-1) ? ", " : "");
 		(void)fprintf(stderr,")\n");
 		(void)fprintf(stderr,"  edge = (");
 		for (id = 0 ; id < test.vars[iv].ndims; id++)
-		  (void)fprintf(stderr,"%d%s",hc[ie].edg[id],
+		  (void)fprintf(stderr,"%d%s",(int)hc[ie].edg[id],
 				(id < test.vars[iv].ndims-1) ? ", " : "");
 		(void)fprintf(stderr,")\n");
 		(void)fprintf(stderr,"  external strides = (");
 		for (id = 0 ; id < test.vars[iv].ndims; id++)
-		  (void)fprintf(stderr,"%d%s",strides[id],
+		  (void)fprintf(stderr,"%d%s",(int)strides[id],
 				(id < test.vars[iv].ndims-1) ? ", " : "");
 		(void)fprintf(stderr,")\n");
 		(void)fprintf(stderr,"  internal basis vector = (");
 		for (id = 0 ; id < test.vars[iv].ndims; id++)
-		  (void)fprintf(stderr,"%d%s",basis[id],
+		  (void)fprintf(stderr,"%d%s",(int)basis[id],
 				(id < test.vars[iv].ndims-1) ? ", " : "");
 		(void)fprintf(stderr,")\n");
 	    } else {
